Factor out GitHub presenter into separate package.
This allows more easily overriding the HTTP client used by the GitHub
presenter for accessing GitHub API.
